Refine crosscheck test oracle coverage
Add the crosscheck/test oracle design and move JDK-compatible named-group coverage into the ordinary public API suites. Replace stale SafeRE API test classes with internal metadata coverage, and remove stale generated-crosscheck disables from executable RE2 corpus tests.

Refs #269
