+++ safere/src/test/java/org/safere/MatcherTest.java
@@ -653,6 +653,114 @@ class MatcherTest {
     }
 
     @Test
+    @DisplayName("toMatchResult() snapshot supports named-group lookup")
+    void toMatchResultNamedGroups() {
+      Pattern p = Pattern.compile("(?<word>\\w+)");
+      Matcher m = p.matcher("hello");
+      assertThat(m.find()).isTrue();
+
+      MatchResult mr = m.toMatchResult();
+      assertThat(mr.namedGroups()).containsEntry("word", 1);
+      assertThat(mr.group("word")).isEqualTo("hello");
+      assertThat(mr.start("word")).isEqualTo(0);
+      assertThat(mr.end("word")).isEqualTo(5);
+    }
+
+    @Test
+    @DisplayName("start(String) and end(String) return named group positions")
+    void namedGroupStartEnd() {
+      Pattern p = Pattern.compile("(?<word>\\w+)@(?<host>\\w+)");
+      Matcher m = p.matcher("user@host");
+      assertThat(m.find()).isTrue();
+      assertThat(m.start("word")).isEqualTo(0);
+      assertThat(m.end("word")).isEqualTo(4);
+      assertThat(m.start("host")).isEqualTo(5);
+      assertThat(m.end("host")).isEqualTo(9);
+    }
+
+    @Test
+    @DisplayName("start(String) throws for unknown name")
+    void namedGroupStartUnknownThrows() {
+      Pattern p = Pattern.compile("(?<word>\\w+)");
+      Matcher m = p.matcher("hello");
+      m.find();
+      assertThatThrownBy(() -> m.start("missing"))
+          .isInstanceOf(IllegalArgumentException.class);
+    }
+
+    @Test
+    @DisplayName("end(String) throws for unknown name")
+    void namedGroupEndUnknownThrows() {
+      Pattern p = Pattern.compile("(?<word>\\w+)");
+      Matcher m = p.matcher("hello");
+      m.find();
+      assertThatThrownBy(() -> m.end("missing"))
+          .isInstanceOf(IllegalArgumentException.class);
+    }
+
+    @Test
+    @DisplayName("named group that did not participate returns null and -1 positions")
+    void nonParticipatingNamedGroup() {
+      Pattern p = Pattern.compile("(?<a>a)|(?<b>b)");
+      Matcher m = p.matcher("b");
+      assertThat(m.find()).isTrue();
+      assertThat(m.group("a")).isNull();
+      assertThat(m.start("a")).isEqualTo(-1);
+      assertThat(m.end("a")).isEqualTo(-1);
+      assertThat(m.group("b")).isEqualTo("b");
+      assertThat(m.start("b")).isEqualTo(0);
+      assertThat(m.end("b")).isEqualTo(1);
+    }
+
+    @Test
+    @DisplayName("namedGroups() returns named groups from pattern")
+    void namedGroupsReturnsMap() {
+      Pattern p = Pattern.compile("(?<user>\\w+)@(?<host>\\w+)");
+      Matcher m = p.matcher("user@host");
+      assertThat(m.namedGroups()).containsEntry("user", 1);
+      assertThat(m.namedGroups()).containsEntry("host", 2);
+    }
+
+    @Test
+    @DisplayName("namedGroups() returns empty map for no named groups")
+    void namedGroupsEmpty() {
+      Pattern p = Pattern.compile("(\\w+)@(\\w+)");
+      Matcher m = p.matcher("user@host");
+      assertThat(m.namedGroups()).isEmpty();
+    }
+
+    @Test
+    @DisplayName("namedGroups() is unmodifiable")
+    void namedGroupsUnmodifiable() {
+      Pattern p = Pattern.compile("(?<name>\\w+)");
+      Matcher m = p.matcher("hello");
+      assertThatThrownBy(() -> m.namedGroups().put("foo", 99))
+          .isInstanceOf(UnsupportedOperationException.class);
+    }
+
+    @Test
+    @DisplayName("namedGroups() returns from MatchResult interface")
+    void namedGroupsFromMatchResult() {
+      Pattern p = Pattern.compile("(?<word>\\w+)");
+      Matcher m = p.matcher("hello");
+      assertThat(m.find()).isTrue();
+      MatchResult result = m.toMatchResult();
+      assertThat(result.namedGroups()).containsEntry("word", 1);
+    }
+
+    @Test
+    @DisplayName("named group methods reject null names")
+    void namedGroupMethodsRejectNullNames() {
+      Pattern p = Pattern.compile("(?<word>\\w+)");
+      Matcher m = p.matcher("hello");
+      assertThat(m.find()).isTrue();
+
+      assertThatThrownBy(() -> m.group((String) null)).isInstanceOf(NullPointerException.class);
+      assertThatThrownBy(() -> m.start((String) null)).isInstanceOf(NullPointerException.class);
+      assertThatThrownBy(() -> m.end((String) null)).isInstanceOf(NullPointerException.class);
+    }
+
+    @Test
     @DisplayName("groupCount() returns number of capturing groups (excluding group 0)")
     void groupCount() {
       Pattern p = Pattern.compile("(a)(b)(c)");
@@ -953,6 +1061,14 @@ class MatcherTest {
       Pattern p = Pattern.compile("(?<word>\\w+)");
       Matcher m = p.matcher("hello world");
       assertThat(m.replaceAll("${word}!")).isEqualTo("hello! world!");
+    }
+
+    @Test
+    @DisplayName("replaceFirst() with named backreference")
+    void replaceFirstWithNamedBackref() {
+      Pattern p = Pattern.compile("(?<word>\\w+)");
+      Matcher m = p.matcher("hello world");
+      assertThat(m.replaceFirst("${word}!")).isEqualTo("hello! world");
     }
 
     @Test
